# Vectorize class Tensor without requiring alignment or padding.