--- a/include/deal.II/base/tensor.h
+++ b/include/deal.II/base/tensor.h
@@ -840,9 +840,9 @@ public:
    *
    * @note This function can also be used in @ref GlossDevice "device" code.
    */
-   DEAL_II_HOST_DEVICE
-    typename numbers::NumberTraits<Number>::real_type
-    norm_square() const;
+  DEAL_II_HOST_DEVICE
+  typename numbers::NumberTraits<Number>::real_type
+  norm_square() const;
 
   /**
    * Fill a vector with all tensor elements.
@@ -2160,17 +2160,15 @@ constexpr DEAL_II_HOST_DEVICE_ALWAYS_INLINE
  * @relatesalso Tensor
  */
 template <int rank, int dim, typename Number, typename OtherNumber>
-constexpr DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
+DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
   Tensor<rank,
          dim,
          typename ProductType<Number,
                               typename EnableIfScalar<OtherNumber>::type>::type>
   operator*(const Tensor<rank, dim, Number> &t, const OtherNumber &factor)
 {
-  // recurse over the base objects
-  Tensor<rank, dim, typename ProductType<Number, OtherNumber>::type> tt;
-  for (unsigned int d = 0; d < dim; ++d)
-    tt[d] = t[d] * factor;
+  Tensor<rank, dim, typename ProductType<Number, OtherNumber>::type> tt(t);
+  tt *= factor;
   return tt;
 }
 
@@ -2188,7 +2186,7 @@ constexpr DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
  * @relatesalso Tensor
  */
 template <int rank, int dim, typename Number, typename OtherNumber>
-DEAL_II_HOST_DEVICE constexpr inline DEAL_II_ALWAYS_INLINE
+DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
   Tensor<rank,
          dim,
          typename ProductType<typename EnableIfScalar<Number>::type,
@@ -2211,28 +2209,15 @@ DEAL_II_HOST_DEVICE constexpr inline DEAL_II_ALWAYS_INLINE
  * @relatesalso Tensor
  */
 template <int rank, int dim, typename Number, typename OtherNumber>
-constexpr DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
+DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
   Tensor<rank,
          dim,
          typename ProductType<Number,
                               typename EnableIfScalar<OtherNumber>::type>::type>
   operator/(const Tensor<rank, dim, Number> &t, const OtherNumber &factor)
 {
-  Tensor<rank, dim, typename ProductType<Number, OtherNumber>::type> tt;
-  if constexpr (std::is_integral<
-                  typename ProductType<Number, OtherNumber>::type>::value)
-    {
-      // recurse over the base objects
-      for (unsigned int d = 0; d < dim; ++d)
-        tt[d] = t[d] / factor;
-    }
-  else
-    {
-      const Number inverse_factor = Number(1.) / factor;
-      // recurse over the base objects
-      for (unsigned int d = 0; d < dim; ++d)
-        tt[d] = t[d] * inverse_factor;
-    }
+  Tensor<rank, dim, typename ProductType<Number, OtherNumber>::type> tt(t);
+  tt /= factor;
   return tt;
 }
 
@@ -2253,10 +2238,7 @@ DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
             const Tensor<rank, dim, OtherNumber> &q)
 {
   Tensor<rank, dim, typename ProductType<Number, OtherNumber>::type> tmp(p);
-
-  for (unsigned int i = 0; i < dim; ++i)
-    tmp[i] += q[i];
-
+  tmp += q;
   return tmp;
 }
 
@@ -2277,10 +2259,7 @@ DEAL_II_HOST_DEVICE inline DEAL_II_ALWAYS_INLINE
             const Tensor<rank, dim, OtherNumber> &q)
 {
   Tensor<rank, dim, typename ProductType<Number, OtherNumber>::type> tmp(p);
-
-  for (unsigned int i = 0; i < dim; ++i)
-    tmp[i] -= q[i];
-
+  tmp -= q;
   return tmp;
 }