--- a/include/deal.II/physics/elasticity/standard_tensors.h
+++ b/include/deal.II/physics/elasticity/standard_tensors.h
@@ -230,8 +230,8 @@ namespace Physics
        * @dealiiHolzapfelA{229,6.83}
        */
       template <typename Number>
-      static DEAL_II_HOST DEAL_II_CONSTEXPR SymmetricTensor<4, dim, Number>
-      Dev_P(const Tensor<2, dim, Number> &F);
+      static DEAL_II_HOST SymmetricTensor<4, dim, Number>
+                          Dev_P(const Tensor<2, dim, Number> &F);
 
       /**
        * Return the transpose of the fourth-order referential deviatoric tensor,
@@ -245,8 +245,8 @@ namespace Physics
        * @f]
        */
       template <typename Number>
-      static DEAL_II_HOST DEAL_II_CONSTEXPR SymmetricTensor<4, dim, Number>
-      Dev_P_T(const Tensor<2, dim, Number> &F);
+      static DEAL_II_HOST SymmetricTensor<4, dim, Number>
+                          Dev_P_T(const Tensor<2, dim, Number> &F);
 
       /** @} */
 
@@ -272,8 +272,8 @@ namespace Physics
        * @dealiiHolzapfelA{228,6.82}
        */
       template <typename Number>
-      static DEAL_II_HOST DEAL_II_CONSTEXPR SymmetricTensor<2, dim, Number>
-      ddet_F_dC(const Tensor<2, dim, Number> &F);
+      static DEAL_II_HOST SymmetricTensor<2, dim, Number>
+                          ddet_F_dC(const Tensor<2, dim, Number> &F);
 
       /** @} */
 
@@ -297,8 +297,8 @@ namespace Physics
        * @dealiiWriggersA{76,3.255}
        */
       template <typename Number>
-      static DEAL_II_HOST DEAL_II_CONSTEXPR SymmetricTensor<4, dim, Number>
-      dC_inv_dC(const Tensor<2, dim, Number> &F);
+      static DEAL_II_HOST SymmetricTensor<4, dim, Number>
+                          dC_inv_dC(const Tensor<2, dim, Number> &F);
 
       /** @} */
     };
@@ -315,8 +315,8 @@ namespace Physics
 
 template <int dim>
 template <typename Number>
-DEAL_II_HOST DEAL_II_CONSTEXPR inline SymmetricTensor<4, dim, Number>
-             Physics::Elasticity::StandardTensors<dim>::Dev_P(
+DEAL_II_HOST inline SymmetricTensor<4, dim, Number>
+Physics::Elasticity::StandardTensors<dim>::Dev_P(
   const Tensor<2, dim, Number> &F)
 {
   // Make things work with AD types
@@ -342,8 +342,8 @@ DEAL_II_HOST DEAL_II_CONSTEXPR inline SymmetricTensor<4, dim, Number>
 
 template <int dim>
 template <typename Number>
-DEAL_II_HOST DEAL_II_CONSTEXPR inline SymmetricTensor<4, dim, Number>
-             Physics::Elasticity::StandardTensors<dim>::Dev_P_T(
+DEAL_II_HOST inline SymmetricTensor<4, dim, Number>
+Physics::Elasticity::StandardTensors<dim>::Dev_P_T(
   const Tensor<2, dim, Number> &F)
 {
   // Make things work with AD types
@@ -370,8 +370,8 @@ DEAL_II_HOST DEAL_II_CONSTEXPR inline SymmetricTensor<4, dim, Number>
 
 template <int dim>
 template <typename Number>
-DEAL_II_HOST DEAL_II_CONSTEXPR SymmetricTensor<2, dim, Number>
-Physics::Elasticity::StandardTensors<dim>::ddet_F_dC(
+DEAL_II_HOST SymmetricTensor<2, dim, Number>
+             Physics::Elasticity::StandardTensors<dim>::ddet_F_dC(
   const Tensor<2, dim, Number> &F)
 {
   return internal::NumberType<Number>::value(0.5 * determinant(F)) *
@@ -382,8 +382,8 @@ Physics::Elasticity::StandardTensors<dim>::ddet_F_dC(
 
 template <int dim>
 template <typename Number>
-DEAL_II_HOST DEAL_II_CONSTEXPR inline SymmetricTensor<4, dim, Number>
-             Physics::Elasticity::StandardTensors<dim>::dC_inv_dC(
+DEAL_II_HOST inline SymmetricTensor<4, dim, Number>
+Physics::Elasticity::StandardTensors<dim>::dC_inv_dC(
   const Tensor<2, dim, Number> &F)
 {
   const SymmetricTensor<2, dim, Number> C_inv =