--- a/tests/base/point_constexpr.cc
+++ b/tests/base/point_constexpr.cc
@@ -20,21 +20,11 @@
 #include "../tests.h"
 
 
-template <int dim>
-constexpr bool val =
-  Point<dim>((Point<dim>::unit_vector(0) + Point<dim>::unit_vector(0)) -
-             2. * Point<dim>::unit_vector(0))[0] == 0.;
-
-
 int
 main()
 {
   initlog();
   deallog << std::setprecision(3);
-
-  static_assert(val<1>);
-  static_assert(val<2>);
-  static_assert(val<3>);
 
   static_assert((Point<1>() / 1.).norm_square() == 0.);
   static_assert(Point<2>{0., 1.}.distance_square(Point<2>{}) == 1.);