--- a/tests/tensors/constexpr_tensor.output
+++ b/tests/tensors/constexpr_tensor.output
@@ -165,5 +165,4 @@ DEAL:double::0.00000 0.00000 0.00000 0.00000 0.00000 0.00000 0.00000 0.00000 0.0
 DEAL::Using Tensor within constexpr functions
 DEAL::15.2000
 DEAL::116.000
-DEAL::Used memory: 72
 DEAL::OK