… tests/tensors/constexpr_symmetric_tensor.cc
… tests/tensors/constexpr_tensor.cc