This is an alternative to #16848, in pursuit of #clean up #16465. In contrast to #16848, where I do `reinterpret_cast`s of the underlying storage to a `VectorizedArray` type (which requires incresing the alignment of `Tensor` and the use of padding elements), here I'm just reconstituting `VectorizedArray` objects every time I need one, hoping the compiler will choose efficient ways of creating a, say, `_m256` from the data that is in memory. 

I'll report timings in a separate comment below.