@bangerth 2024-04-06
For the same benchmark as shown in #16848 (a marginally modified version of step-22), I get these timings:
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      66.3s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |      2.24s |       3.4% |
| compute preconditioner          |         1 |      1.17s |       1.8% |
| output results                  |         1 |    0.0957s |      0.14% |
| refine mesh                     |         1 |    0.0425s |         0% |
| setup                           |         1 |     0.653s |      0.98% |
| solve                           |         1 |      62.1s |        94% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      66.8s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |      2.27s |       3.4% |
| compute preconditioner          |         1 |      1.18s |       1.8% |
| output results                  |         1 |    0.0958s |      0.14% |
| refine mesh                     |         1 |    0.0413s |         0% |
| setup                           |         1 |     0.662s |      0.99% |
| solve                           |         1 |      62.5s |        94% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      66.1s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |      2.27s |       3.4% |
| compute preconditioner          |         1 |      1.17s |       1.8% |
| output results                  |         1 |    0.0954s |      0.14% |
| refine mesh                     |         1 |    0.0417s |         0% |
| setup                           |         1 |     0.656s |      0.99% |
| solve                           |         1 |      61.9s |        94% |
+---------------------------------+-----------+------------+------------+
```
So the best times are for "assemble" and "refine mesh", respectively:

* Without this patch: 2.00 and 0.0384 seconds (taken from #16848)
* With this patch: 2.24 and 0.0413 seconds

In other words, assembly is *even slower* than the other approach; mesh refinement (primarily calling the Kelly estimator) is slightly faster than in the other patch, but still slower than current master.