@bangerth 2024-04-06
I'll leave this open for a couple more days as well, in case someone wanted to play with it, but I think that at the end of the day the idea of using vector instructions for `class Tensor` just did not work out. I don't plan on spending any more time on this.